requirements.txt
asgiref==3.8.1
backports.tarfile==1.1.1
boto3==1.34.113
botocore==1.34.113
Brotli==1.1.0
cffi==1.16.0
click==8.1.7
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
dj-database-url==2.1.0
Django==5.0.6
django-admin-sortable2==2.2.1
django-storages==1.14.3
fonttools==4.52.4
gunicorn==22.0.0
h11==0.14.0
importlib_metadata==7.1.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jeepney==0.8.0
jmespath==1.0.1
keyring==25.2.1
kiwisolver==1.4.5
matplotlib==3.9.0
more-itertools==10.2.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pyparsing==3.1.2
pysam==0.22.1
python-dateutil==2.9.0.post0
RUST==1.3.1
s3transfer==0.10.1
SecretStorage==3.3.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
urllib3==2.2.1
uvicorn==0.29.0
whitenoise==6.6.0
zipp==3.18.2
